In accordance with client update "Buttons"
Added a toString() for the Actor, VisibleObject, and GameObject classes.
Added a constructor for the TerrainObject class.

Made updates to the ClientListener class in accordance with the changes to the SLEClient. Primarily focused around the client's new ability to have on screen buttons but will apply to other HUD elements in the future.